Read ME!
Sugar Run is inspired from games like Sugarrush and ZigZag wich are easy and fun to play. My game is basically a Jump and Run game in 3D made for Smartphones as well as computers.
You only need to press one button to move the gamer and need to think fast to manage all levels.